Fixed issues related to conventional/primitive lattice detection
Plus plotting Brillouin zones should be more robust now for
ordering face vertices correctly. The algorithm is now
deterministic and will always work under the constraint that
faces are formed by convex polygons, and may still work for some
types of partially concave polygons.

Other minor changes to try and appease the MSVC compiler by
removing warnings.

+++ src/bz.cpp
@@ -271,7 +271,7 @@ bool BrillouinZone::moveinto(const LQVec<double>& Q, LQVec<double>& q, LQVec<int
   LQVec<int> taui;
   ArrayVector<double> q_dot_facenrm;
   ArrayVector<int> Nhkl;
-  int maxat = 0;
+  size_t maxat = 0;
   int maxnm = 0;
   size_t count =0;
   for (size_t i=0; i<Qsl.size(); i++){
